+++ src/main/java/altamirano/hernandez/proyectogastos_springboot_angular/repositories/IProveedorRepository.java
@@ -3,5 +3,5 @@ package altamirano.hernandez.proyectogastos_springboot_angular.repositories;
 import altamirano.hernandez.proyectogastos_springboot_angular.models.Proveedor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IProveedorRepository extends CrudRepository<Proveedor, Long> {
+public interface IProveedorRepository extends CrudRepository<Proveedor, Integer> {
 }
