+++ src/main/java/altamirano/hernandez/proyectogastos_springboot_angular/services/interfaces/IProveedorService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface IProveedorService {
     public abstract List<Proveedor> findAll();
-    public abstract Optional<Proveedor> findById(Long id);
+    public abstract Optional<Proveedor> findById(int id);
     public abstract void save(Proveedor proveedor);
-    public abstract void deleteById(Long id);
+    public abstract void deleteById(int id);
 }
